site_agent 파일 정리
